% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_add_rule}
\alias{CTS_add_rule}
\title{Add Rule to Element}
\usage{
CTS_add_rule(state, element)
}
\arguments{
\item{state}{CTS state from \code{CTS_fetch_state()}}

\item{element}{Element list from \code{CTS_fetch_current_element()}}
}
\value{
Element with the results of adding the rule. The \code{rares} list
element can be used to determine the exit status of the function.
\itemize{
\item{RULE_IS_GOOD}  If true it indicates that the pieces of the rule from
the UI check out.
\item{RULE_UPDATED}  If RULE_IS_GOOD and RULE_UPDATED is true then a
previous rule definition was overwritten. If RULE_IS_GOOD is TRUE and
RULE_UPDATED is FALSE then a new rule was added.
\item{notify_text}  Text for notify message
\item{notify_id}    Notification ID
\item{notify_type}  Notification type
\item{msgs}         Vector of messages.
}
}
\description{
Takes the ui elements in the state and element and attempts to add a rule.
}
\details{
This depends on the following UI values in the state and element
\itemize{
\item{}  \code{state[["CTS"]][["ui"]][["rule_name"]]}
\item{}  \code{state[["CTS"]][["ui"]][["rule_condition"]]}
\item{}  \code{state[["CTS"]][["ui"]][["rule_type"]]}
\itemize{
\item{}  For rule type "dose"
\itemize{
\item{}  \code{state[["CTS"]][["ui"]][["action_dosing_state"]]}
\item{}  \code{state[["CTS"]][["ui"]][["action_dosing_values"]]}
\item{}  \code{state[["CTS"]][["ui"]][["action_dosing_times"]]}
\item{}  \code{state[["CTS"]][["ui"]][["action_dosing_durations"]]}
}
}
\itemize{
\item{}  For rule type "set state"
\itemize{
\item{}  \code{state[["CTS"]][["ui"]][["action_set_state_state"]]}
\item{}  \code{state[["CTS"]][["ui"]][["action_set_state_values"]]}
}
}
\itemize{
\item{}  For rule type "manual code"
\itemize{
\item{}  \code{state[["CTS"]][["ui"]][["action_manual_code"]]}
}
}
}
}
\examples{
# For more information see the Clinical Trial Simulation vignette:
# https://ruminate.ubiquity.tools/articles/clinical_trial_simulation.html

# None of this will work if rxode2 isn't installed:
library(formods)
if(is_installed("rxode2")){

sess_res = CTS_test_mksession(session=list(), full_session=FALSE)
state = sess_res$state
session = sess_res$session
input   = sess_res$input

# Fetch a list of the current element
element = CTS_fetch_current_element(state)

# You can modify the element
element[["element_name"]] = "A more descriptive name"

# Simulating the current element
element = CTS_simulate_element(state, element)

# You can now place element back in the state
state = CTS_set_current_element(state, element)

# This will extract the code for the current module
code = CTS_fetch_code(state)
code

# This will update the checksum of the module state
state = CTS_update_checksum(state)


# CTS_add_covariate
# CTS_add_rule
# CTS_change_source_model
# CTS_fetch_code
# CTS_plot_element


# Creates a new empty element
state = CTS_new_element(state)

# Delete the current element
state = CTS_del_current_element(state)
}
}
