% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_fetch_code}
\alias{CTS_fetch_code}
\title{Fetch Module Code}
\usage{
CTS_fetch_code(state)
}
\arguments{
\item{state}{CTS state from \code{CTS_fetch_state()}}
}
\value{
Character object vector with the lines of code
}
\description{
Fetches the code to generate results seen in the app
}
\examples{
# We need a module state:
sess_res = CTS_test_mksession(session=list())
state = sess_res$state

code = CTS_fetch_code(state)

cat(code)
}
