% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_fetch_ds}
\alias{CTS_fetch_ds}
\title{Fetch Clinical Trial Simulator Module Datasets}
\usage{
CTS_fetch_ds(state)
}
\arguments{
\item{state}{CTS state from \code{CTS_fetch_state()}}
}
\value{
Character object vector with the lines of code

list containing the following elements
\itemize{
\item{isgood:}    Return status of the function.
\item{hasds:}     Boolean indicator if the module has any datasets
\item{msgs:}      Messages to be passed back to the user.
\item{ds:}        List with datasets. Each list element has the name of
the R-object for that dataset. Each element has the following structure:
\itemize{
\item{label: Text label for the dataset}
\item{MOD_TYPE: Short name for the type of module.}
\item{id: module ID}
\item{DS: Dataframe containing the actual dataset.}
\item{DSMETA: Metadata describing DS}
\item{code: Complete code to build dataset.}
\item{checksum: Module checksum.}
\item{DSchecksum: Dataset checksum.}
}
}
}
\description{
Fetches the datasets produced by the module. For each cohort
this will be the simulation timecourse and the event table
}
\examples{
# We need a module state:
sess_res = CTS_test_mksession(session=list(), full_session=FALSE)
state = sess_res$state

ds = CTS_fetch_ds(state)

ds
}
