% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_mk_preload}
\alias{CTS_mk_preload}
\title{Make List of Current CTS State}
\usage{
CTS_mk_preload(state)
}
\arguments{
\item{state}{CTS state object}
}
\value{
list with the following elements
\itemize{
\item{isgood:}       Boolean indicating the exit status of the function.
\item{msgs:}         Messages to be passed back to the user.
\item{yaml_list:}    Lists with preload components.
}
}
\description{
Reads in the app state from yaml files.
}
\examples{
\donttest{
sess_res = CTS_test_mksession()
state = sess_res$state
res = CTS_mk_preload(state)
}
}
