% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_sim_isgood}
\alias{CTS_sim_isgood}
\title{Checks Simulation in Element for Goodness}
\usage{
CTS_sim_isgood(state, element)
}
\arguments{
\item{state}{CTS state from \code{CTS_fetch_state()}}

\item{element}{Element list from \code{CTS_fetch_current_element()}}
}
\value{
List with the following elements:
\itemize{
\item{isgood:} Boolean object indicating if the file was successfully loaded.
\item{msgs:} Text description of failure.
}
}
\description{
Takes the supplied element and determines if the underlying
simulation is in a good state or not.
}
