% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_test_mksession}
\alias{CTS_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
CTS_test_mksession(session = list(), full = FALSE)
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}

\item{full}{Boolean indicating if the full test session should be created
(\code{TRUE}) or a minimal test session should be created (\code{FALSE},
default)}
}
\value{
The CTS portion of the \code{all_sess_res} returned from \code{\link{FM_app_preload}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
\donttest{
session = shiny::MockShinySession$new()
sess_res = CTS_test_mksession(session=session)
}
}
\seealso{
\code{\link{FM_app_preload}}
}
