% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_update_checksum}
\alias{CTS_update_checksum}
\title{Updates CTS Module Checksum}
\usage{
CTS_update_checksum(state)
}
\arguments{
\item{state}{CTS state from \code{CTS_fetch_state()}}
}
\value{
CTS state object with the checksum updated
}
\description{
Takes a CTS state and updates the checksum used to trigger
downstream updates
}
\examples{
# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = CTS_test_mksession(session=list())
session = sess_res$session
input   = sess_res$input

# We also need a state variable
state = sess_res$state

state = CTS_update_checksum(state)
}
