% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_rules.R
\name{fetch_rxtc}
\alias{fetch_rxtc}
\title{Extracts Timecourse and Merges Covariates}
\usage{
fetch_rxtc(rx_details, sim)

fetch_rxtc(rx_details, sim)
}
\arguments{
\item{rx_details}{Output of \code{fetch_rxinfo()}}

\item{sim}{output of \code{rxSolve()}}
}
\value{
Dataframe of the simulated time course.

Dataframe of the simulated time course.
}
\description{
Takes the output of \code{rxSolve()} and merges in any missing
covariates that are present in params but not in sim

Takes the output of \code{rxSolve()} and merges in any missing
covariates that are present in params but not in sim
}
