% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Server.R
\name{MB_fetch_code}
\alias{MB_fetch_code}
\title{Fetch Module Code}
\usage{
MB_fetch_code(state)
}
\arguments{
\item{state}{MB state from \code{MB_fetch_state()}}
}
\value{
Character object vector with the lines of code
}
\description{
Fetches the code to generate results seen in the app
}
\examples{
# We need a module state:
sess_res = MB_test_mksession(session=list(), full_session=FALSE)
state = sess_res$state

code = MB_fetch_code(state)

cat(code)
}
