% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Server.R
\name{MB_mk_preload}
\alias{MB_mk_preload}
\title{Make List of Current MB State}
\usage{
MB_mk_preload(state)
}
\arguments{
\item{state}{MB state object}
}
\value{
list with the following elements
\itemize{
\item{isgood:}       Boolean indicating the exit status of the function.
\item{msgs:}         Messages to be passed back to the user.
\item{yaml_list:}    Lists with preload components.
}
}
\description{
Reads in the app state from yaml files.
}
\examples{
res = MB_mk_preload(state)
}
