% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Server.R
\name{MB_mod_ok}
\alias{MB_mod_ok}
\title{Makes Sure Suggests are Installed}
\usage{
MB_mod_ok(pkgs = c("rxode2", "nonmem2rx", "nlmixr2lib"))
}
\arguments{
\item{pkgs}{List of suggested packages (default "rxode2", "nonmem2rx" and "nlmixr2lib")}
}
\value{
Boolean value indicating if the module is OK
}
\description{
Simple check to make sure the suggested packages are installed.
}
\examples{
MB_mod_ok()
}
