% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Server.R
\name{MB_update_checksum}
\alias{MB_update_checksum}
\title{Update MB Module Checksum}
\usage{
MB_update_checksum(state)
}
\arguments{
\item{state}{MB state from \code{MB_fetch_state()}}
}
\value{
MB state object with the checksum updated
}
\description{
Takes a MB state and updates the checksum used to trigger
downstream updates
}
\examples{
# Within shiny both session and input variables will exist,
# this creates examples here for testing purposes:
sess_res = MB_test_mksession(session=list())
session = sess_res$session
input   = sess_res$input

# We also need a state variable
state = sess_res$state

state = MB_update_checksum(state)
}
