% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{mk_table_nca_params}
\alias{mk_table_nca_params}
\title{Title}
\usage{
mk_table_nca_params(
  nca_res,
  type = "individual",
  grouping = "interval",
  not_calc = "NC",
  obnd = NULL,
  nps = NULL,
  infinity = list(text = "∞", md = "∞"),
  digits = 3,
  max_row = NULL,
  max_col = NULL
)
}
\arguments{
\item{nca_res}{Output of PKNCA.}

\item{type}{Type of table to generate. Can be either \code{"individual"} or \code{"summary"]}.}

\item{grouping}{How to group columns in tables. Can be either \code{"interval"} or \code{"parameter"]}.}

\item{obnd}{onbrand reporting object.}

\item{nps}{NCA parameter summary table with the following columns.
\itemize{
\item{parameter:}      PKNCA Paramter name.
\item{text:}           Name used in text output.
\item{md:}             Name used markdown output.
\item{latex:}          Name used in latex output.
\item{description:}    Verbose textual description of the parameter.
}}

\item{digits}{Number of significant figures to report (set to \code{NULL}
to disable rounding)}

\item{max_col}{Maximum number of columns to have on a page. Spillover will
be wrapped to multiple pages.}
}
\value{
list containing the following elements
\itemize{
\item{raw_nca:} Raw PKNCA output.
\item{isgood:}  Boolean indicating the exit status of the function.
\item{msgs:}    Vector of text messages describing any errors that were found.
\item{tables:}  Named list of tables. Each list element is of the output
format from \code{build_span()}.
}
}
\description{
Description
}
