% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_fetch_code}
\alias{NCA_fetch_code}
\title{Fetch Module Code}
\usage{
NCA_fetch_code(state)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}}
}
\value{
Character object vector with the lines of code
}
\description{
Fetches the code to generate results seen in the app
}
