% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_fetch_current_ana}
\alias{NCA_fetch_current_ana}
\title{Fetches Current Analysis}
\usage{
NCA_fetch_current_ana(state)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}}
}
\value{
List containing the details of the current analysis. The structure
of this list is the same as the structure of \code{state$NCA$anas} in the output of
\code{NCA_fetch_state()}.
}
\description{
Takes an NCA state and returns the current active analysis
}
\examples{
# We need a state object to use below
sess_res = NCA_test_mksession(session=list())
state = sess_res$state
current_ana = NCA_fetch_current_ana(state)
}
