% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_fetch_data_format}
\alias{NCA_fetch_data_format}
\title{Fetches Details About Data Requirements}
\usage{
NCA_fetch_data_format(
  MOD_yaml_file = system.file(package = "ruminate", "templates", "NCA.yaml")
)
}
\arguments{
\item{MOD_yaml_file}{Module configuration file with MC as main section.}
}
\value{
List with details about the data formats
}
\description{
Use this to get information about data formats.
}
\examples{
NCA_fetch_data_format()
}
