% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_fetch_PKNCA_meta}
\alias{NCA_fetch_PKNCA_meta}
\title{Fetches PKNCA Metadata}
\usage{
NCA_fetch_PKNCA_meta()
}
\value{
Dataframe containing PKCNA metadata for NCA parameters.
}
\description{
Compiles Metadata from PKNCA
}
\examples{
PKNCA_meta = NCA_fetch_PKNCA_meta()
utils::head(PKNCA_meta)
}
