% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_init_state}
\alias{NCA_init_state}
\title{Initialize NCA Module State}
\usage{
NCA_init_state(FM_yaml_file, MOD_yaml_file, id)
}
\arguments{
\item{FM_yaml_file}{App configuration file with FM as main section.}

\item{MOD_yaml_file}{Module configuration file with MC as main section.}

\item{id}{ID string for the module.}
}
\value{
list containing an empty NCA state
}
\description{
Creates a list of the initialized module state
}
