% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_new_ana}
\alias{NCA_new_ana}
\title{Initialize New Analysis}
\usage{
NCA_new_ana(state)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}}
}
\value{
NCA state object containing a new empty analysis and that analysis
is set as the current active analyisis
}
\description{
Creates a new NCA analysis in an NCA module
}
