% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_reset_manual_flags}
\alias{NCA_reset_manual_flags}
\title{Reset Manual Flags}
\usage{
NCA_reset_manual_flags(state)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}}
}
\value{
state with manual flags reset
}
\description{
Resets the manual flags for the current analysis
}
\examples{
# We need a state object to use below
sess_res = NCA_test_mksession()
state = sess_res$state
state = NCA_reset_manual_flags(state)
}
