% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_set_current_ana}
\alias{NCA_set_current_ana}
\title{Sets Current Analysis}
\usage{
NCA_set_current_ana(state, ana)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}}

\item{ana}{Analysis list from \code{NCA_fetch_current_ana}}
}
\value{
State with the current analysis updated
}
\description{
Takes an NCA state and an analysis list and sets that figure list
as the value for the active figure
}
