% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{run_nca}
\alias{run_nca}
\title{Runs NCA for the Current Analysis}
\usage{
run_nca(state)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}}
}
\value{
List with the following components:
\itemize{
\item{isgood:}    Return status of the function.
\item{msgs:}      Error messages if any issues were encountered.
\item{nca_res:}   PKNCA results if run was successful.
}
}
\description{
Takes the current state and runs the current analysis in that
state.
}
