% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_add_covariate}
\alias{CTS_add_covariate}
\title{Add Covariate to Elemetnt}
\usage{
CTS_add_covariate(state, element)
}
\arguments{
\item{state}{CTS state from \code{CTS_fetch_state()}}

\item{element}{Element list from \code{CTS_fetch_current_element()}}
}
\value{
Element with the results of adding the rule. The \code{cares} list
element can be used to determine the exit status of the function.
\itemize{
\item{RULE_IS_GOOD}  If true it indicates that the pieces of the rule from
the UI check out.
\item{RULE_UPDATED}  If RULE_IS_GOOD and RULE_UPDATED is true then a
previous rule definition was overwritten. If RULE_IS_GOOD is TRUE and
RULE_UPDATED is FALSE then a new rule was added.
\item{notify_text}  Text for notify message
\item{notify_id}    Notification ID
\item{notify_type}  Notification type
\item{msgs}         Vector of messages.
}
}
\description{
Takes the ui elements in the module state and processes the covariate elements for addition.
}
\details{
This depends on the following UI values in the state
\itemize{
\item{}  state[\link{"CTS"}][\link{"ui"}][\link{"covariate_value"}]
\item{}  state[\link{"CTS"}][\link{"ui"}][\link{"covariate_type_selected"}]
\item{}  state[\link{"CTS"}][\link{"ui"}][\link{"selected_covariate"}]
}
}
\examples{
# For more information see the Clinical Trial Simulation vignette:
# https://ruminate.ubiquity.tools/articles/clinical_trial_simulation.html

# None of this will work if rxode2 isn't installed:
if(is_installed("rxode2")){

sess_res = CTS_test_mksession(session=list(), full_session=FALSE)
state = sess_res$state
session = sess_res$session
input   = sess_res$input

# Creates a new empty element
state = CTS_new_element(state)

# Delete the current element
state = CTS_del_current_element(state)

# Fetch a list of the current element
element = CTS_fetch_current_element(state)

# You can modify the element
element[["element_name"]] = "A more descriptive name"

# Simulating the current element
element = CTS_simulate_element(state, element)

# You can now place element back in the state
state = CTS_set_current_element(state, element)

# This will extract the code for the current module
code = CTS_fetch_code(state)
code


# This will update the checksum of the module state
state = CTS_update_checksum(state)

#'@example inst/test_apps/CTS_funcs.R
}
}
