% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTS_Server.R
\name{CTS_test_mksession}
\alias{CTS_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
CTS_test_mksession(session = list())
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
The CTS portion of the \code{all_sess_res} returned from \code{\link{ASM_set_app_state}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
\donttest{
sess_res = CTS_test_mksession()
}
}
\seealso{
\code{\link{ASM_set_app_state}}
}
