% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Server.R
\name{MB_test_catalog}
\alias{MB_test_catalog}
\title{Tests the Model Catalog}
\usage{
MB_test_catalog(state, as_cran = FALSE, verbose = TRUE)
}
\arguments{
\item{state}{MB state from \code{MB_fetch_state()}}

\item{as_cran}{Boolean to indicate if you're running this on CRAN}

\item{verbose}{Boolean to indicate if messages should be displayed.}
}
\value{
List with the following attributes:
}
\description{
Reads in models in the catalog and attempts to build them.
}
