% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MB_Server.R
\name{MB_test_mksession}
\alias{MB_test_mksession}
\title{Populate Session Data for Module Testing}
\usage{
MB_test_mksession(session = list())
}
\arguments{
\item{session}{Shiny session variable (in app) or a list (outside of app)}
}
\value{
The MB portion of the \code{all_sess_res} returned from
\code{\link{FM_app_preload}}
}
\description{
Populates the supplied session variable for testing.
}
\examples{
sess_res = MB_test_mksession()
}
