% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{mk_figure_ind_obs}
\alias{mk_figure_ind_obs}
\title{Creates Figures of Individual Observations from PKNCA Result}
\usage{
mk_figure_ind_obs(
  nca_res,
  OBS_LAB = "Concentration ===CONCUNITS===",
  TIME_LAB = "Time ===TIMEUNITS===",
  OBS_STRING = "Observation",
  BLQ_STRING = "BLQ",
  NA_STRING = "Missing",
  log_scale = TRUE,
  scales = "fixed",
  nfrows = 4,
  nfcols = 3
)
}
\arguments{
\item{nca_res}{Output of PKNCA.}

\item{nfrows}{Number of facet rows per page.}

\item{nfcols}{Number of facet cols per page.}
}
\value{
list containing the following elements
\itemize{
}
}
\description{
Takes the output of PKNCA and creates \code{ggplot} figures faceted
by subject id highlighting of certain NCA aspects (e.g. points used for half-life)
}
