% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{mk_table_ind_obs}
\alias{mk_table_ind_obs}
\title{Creates Tables of Individual Observations from PKNCA Result}
\usage{
mk_table_ind_obs(
  nca_res,
  not_sampled = "NS",
  blq = "BLQ",
  digits = 3,
  max_col = 9,
  rows_by = "time"
)
}
\arguments{
\item{nca_res}{Output of PKNCA.}

\item{not_sampled}{Character string to use for missing data when pivoting.}

\item{digits}{Number of significant figures to report (set to \code{NULL}
to disable rounding)}

\item{max_col}{Maximum number of columns to have on a page. Spillover will
be wrapped to multiple pages.}

\item{rows_by}{Can be either "time" or "id". If it is "time", there will be a
column for time and separate column for each subject ID. If rows_by is set
to "id" there will be a column for ID and a column for each individual time.}
}
\value{
list containing the following elements
\itemize{
}
}
\description{
Takes the output of PKNCA and creates a tabular view of the
individual observation data. This can be spread out of over several tables
(pages) if necessary.
}
