% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{mk_table_nca_params}
\alias{mk_table_nca_params}
\title{Title}
\usage{
mk_table_nca_params(
  nca_res,
  type = "individual",
  nca_params = NULL,
  nps = NULL,
  infinity = list(text = "inf", md = "∞"),
  digits = 3,
  max_col = 9
)
}
\arguments{
\item{nca_res}{Output of PKNCA.}

\item{type}{Type of table to generate. Can be either \code{"individual"} or \code{"summary"]}.}

\item{nps}{NCA parameter summary table with the following columns.
\itemize{
\item{parameter:}      PKNCA Paramter name.
\item{text:}           Name used in text output.
\item{md:}             Name used markdown output.
\item{latex:}          Name used in latex output.
\item{description:}    Verbose textual description of the parameter.
}}

\item{digits}{Number of significant figures to report (set to \code{NULL}
to disable rounding)}

\item{max_col}{Maximum number of columns to have on a page. Spillover will
be wrapped to multiple pages.}
}
\value{
list containing the following elements
\itemize{
}
}
\description{
Description
}
