% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{nca_builder}
\alias{nca_builder}
\title{Builds NCA Code from ui Elements}
\usage{
nca_builder(state)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}
JMH update the return list below}
}
\value{
list containing the following elements
\itemize{
\item{isgood:}           Return status of the function.
\item{msgs:}             Messages to be passed back to the user.
\item{code_previous:}    Code to generate the dataset.
\item{code_ana_only:}    Code for the analysis.
\item{code_fg_ind_obs:}  Code to generate figure(s) of individual observations
\item{code_tb_ind_obs:}  Code to generate table(s) of individual observations
\item{code:}             Complete code to run the analysis.
\item{obj:}              List with names of R objects used in then generated code.
}
}
\description{
Takes the current analysis in the state object and creates the
code to run the analysis
}
