% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_fetch_data_format}
\alias{NCA_fetch_data_format}
\title{Fetches Details About Data Requirements}
\usage{
NCA_fetch_data_format()
}
\value{
list with details about the data formats
}
\description{
Use this to get information about data formats.
}
\examples{
FM_fetch_current_mods()
}
