% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_fetch_PKNCA_meta}
\alias{NCA_fetch_PKNCA_meta}
\title{Fetches PKNCA Metadata}
\usage{
NCA_fetch_PKNCA_meta()
}
\value{
List containing data frames of PKNCA meta data
}
\description{
Compiles Metadata from PKNCA
}
