% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_find_col}
\alias{NCA_find_col}
\title{Determines Default Column Name}
\usage{
NCA_find_col(
  curr_ana = NULL,
  curr_ui = NULL,
  patterns = NULL,
  dscols,
  null_ok = FALSE
)
}
\arguments{
\item{curr_ana}{Current value in the analysis}

\item{curr_ui}{Current value in UI}

\item{patterns}{List of regular expression patterns to consider.}

\item{dscols}{Columns from the dataset.}

\item{null_ok}{Logical value indicating if a null result (nothing found) is
OK (default: \code{FALSE})}
}
\value{
Name of column found based on the rules above.
}
\description{
Based on the current analysis, value from the UI, an optional
list of patterns to search, and column names from a dataset this function
tries to find a default value for a column in the analysis (e.g. subject id,
dose, concentration, etc).

Generally the following is done:
\itemize{
\item If curr_ui has a non-NULL, non-"" value it is compared to dscols. If it
is found there that value is returned.
\item If not then the patterns are considered. If the patterns from the YAML
file are not NULL they are compared sequentially to the columns names.
The first match found is returned.
\item If nothing is found then the first value of dscols is returned.
}
}
