% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_find_col}
\alias{NCA_find_col}
\title{Determines Default Column Name}
\usage{
NCA_find_col(
  curr_ana = NULL,
  curr_ui = NULL,
  patterns = NULL,
  dscols,
  excol = NULL,
  null_ok = FALSE
)
}
\arguments{
\item{curr_ana}{Current value in the analysis}

\item{curr_ui}{Current value in UI}

\item{patterns}{List of regular expression patterns to consider.}

\item{dscols}{Columns from the dataset.}

\item{excol}{Columns to exclude from consideration (default: \code{NULL}).}

\item{null_ok}{Logical value indicating if a null result (nothing found) is
OK (default: \code{FALSE})}
}
\value{
Value List containing the details of the current analysis. The structure
of this list is the same as the structure of \code{state$NCA$anas} in the output of
\code{ANA_fetch_state()}.
}
\description{
Based on the current analysis, value from the UI, an optional list of
patterns to search, column names from a dataset, and an optional list of
column names to exclude it tries to find a default value for a column in the
analysis (e.g. dose, concentration, etc).
\itemize{
\item Excluded columns are removed from dscols
\item If curr_ui has a none NULL non "" value it is compared to dscols. If it
is found there that value is returned.
\item If not then the patterns are considered. IF they are not null they are
compared sequentially to the columns names. The first match found is
returned.
\item If nothing is found then the first value of dscols is returned.
}
}
