% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\name{NCA_load_scenario}
\alias{NCA_load_scenario}
\title{Loads Pre-Defined Scenario}
\usage{
NCA_load_scenario(state, ana_scenario)
}
\arguments{
\item{state}{NCA state from \code{NCA_fetch_state()}}

\item{ana_scenario}{Short name of the analysis scenario to load from the config file.}
}
\value{
NCA state object with the scenario loaded and relevant notifications
set.
}
\description{
Loads a pre-defined analysis scneario from the NCA YAML config
file.
}
