% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCA_Server.R
\docType{package}
\name{ruminate}
\alias{ruminate}
\alias{ruminate-package}
\title{ruminate: Shiny app and module to facilitate pharamacometrics analysis}
\usage{
ruminate()
}
\value{
Nothing.
\itemize{
}
}
\description{
This is done by creating a Shiny interface to different tools for data
transformation (\code{dplyr} and \code{tidyr}), plotting (\code{ggplot2}), and
noncompartmental analysis (\code{PKNCA}). These results can be reported in Excel,
Word or PowerPoint. The state of the app can be saved and loaded at a later
date. When saved, a script is generated to reproduce the different actions in
the Shiny interface.

Runs the pharmacometrics ruminate app.
}
\seealso{
\url{https://ruminate.ubiquity.tools/}
}
\author{
\strong{Maintainer}: John Harrold \email{john.m.harrold@gmail.com} (\href{https://orcid.org/0000-0003-2052-4373}{ORCID})

}
